Changed enums to inherit from int.

+++ src/MelloSilveiraTools.Core/Models/StatusCode.cs
@@ -1,6 +1,6 @@
 ﻿namespace MelloSilveiraTools.Core.Models;
 
-public enum StatusCode : short
+public enum StatusCode : int
 {
     OK = 200,
     Created = 201,
